Fix unit tests when run in Ireland
`new Date(...)` without specifying a timezone appears to default to the user's timezone, so 10am (specified in the test input) equals 7am (specified in the test output)... _if_ the user lives in a UTC+0300 country. If these tests are run from a different country, the test fails. Setting an explicit timezone on the input means that tests should now pass no matter what country they are run in :)